Extract part of auto trading logic
